Give ordered native rules processing one authority
